consolidating database-related tests
